# Triage-app/backend/docs/swagger.yaml
basePath: /
definitions:
...
  services.Assessment:
    properties:
      anatomy_id:
        type: string
      assessment_id:
        type: string
      assessment_type:
        type: string
      completion_percentage:
        type: number
      end_time:
        type: string
      start_time:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  services.ChatMessage:
    properties:
      response:
        description: AI response (optional)
        type: string
      user:
        type: string
    type: object
  services.ChatRequest:
    properties:
      chat_history:
        items:
          $ref: '#/definitions/services.ChatMessage'
        type: array
    type: object
  services.ChatResponse:
    properties:
      response:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for AI-Bot-DeeCogs
  title: AI-Bot-DeeCogs API
  version: "1.0"
paths:
  /assessments:
    post:
      consumes:
      - application/json
      description: Creates a new assessment for the user
      parameters:
      - description: Assessment Data
        in: body
        name: assessment
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.Assessment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a new assessment
      tags:
      - Assessments
...
  /assessments/{id}:
    get:
      description: Retrieves the details of a specific assessment
      parameters:
      - description: Assessment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.Assessment'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get assessment details
      tags:
      - Assessments
...
  /assessments/{id}/chat:
    post:
      consumes:
      - application/json
      description: Sends chat history to an AI model for physiotherapy assessment
      parameters:
      - description: Assessment ID
        in: path
        name: id
        required: true
        type: string
      - description: Chat History
        in: body
        name: chat_body
        required: true
        schema:
          $ref: '#/definitions/services.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.ChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send chat history for AI response
      tags:
      - Assessments
...
  /assessments/{id}/status:
    post:
      consumes:
      - application/json
      description: Updates the status of an assessment
      parameters:
      - description: Assessment ID
        in: path
        name: id
        required: true
        type: string
      - description: Status Update
        in: body
        name: status
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update assessment status
      tags:
      - Assessments
  /auth/loginuser:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password
      parameters:
      - description: User Credentials
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
...
swagger: "2.0"